e2e-integration-test/requirements.txt
pytest==7.3.1
requests==2.28.2
python-dotenv==1.0.0
pytest-html==3.2.0
